FIX: 데이터 삽입이 안되고 꺼져버리는 현상 고침. 일정이 2개보다 적으면 공유버튼 안눌리게 조건 추가

+++ app/src/main/java/com/example/mobilecalendar/roomdb/Alarm.kt
@@ -18,7 +18,7 @@ import java.time.LocalTime
 )
 
 data class Alarm(
-    val scheduleId: Int,
+    val scheduleId: Long,
     val title: String,
     val message: String,
     val time: LocalTime,
